Refactor: Complete JPA migration and service layer updates

+++ src/main/java/com/blog/Repository/PostRepository.java
@@ -1,11 +1,6 @@
 package com.blog.Repository;
 
 import com.blog.Model.Post;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
-
-public interface PostRepository extends JpaRepository<Post, Long> {
-
-}
+public interface PostRepository extends JpaRepository<Post, Integer> { }
